[change] move site logic and attributes to separate dir
