+++ ata_pipeline0/fetch_events.py
@@ -10,7 +10,7 @@ from mypy_boto3_s3.service_resource import ObjectSummary, S3ServiceResource
 from mypy_boto3_s3.type_defs import GetObjectOutputTypeDef
 
 from ata_pipeline0.helpers.logging import logging
-from ata_pipeline0.helpers.site import SiteName
+from ata_pipeline0.site.names import SiteName
 
 logger = logging.getLogger(__name__)
